add strict option, off by default
Previously any missing resources would cause the inliner to halt and
throw an error. This behavior could be considered too safe since a web
page will just render what is available. This update changes the default
behavior to just log a warning when a resource is missing, but adds an
option `strict: true` to use the behavior where the inliner halts on a
missing resource.